feat: refactor admin settings page and system management components - Add new system settings and status components - Update navbar to include system management link - Remove deprecated page settings component - Update setup page to ensure consistent branding
