--- services/frontend/.env
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_API_URL="http://localhost:8080"
